Refactored Content.category into a foreign model ContentCategory
There are dummy methods #category= and #category on the Content model
which *should* help make this a smoother transition to the new
interface. However, this might mean that the migration doesn't roll-back
correctly. In any case, this commit involves a very long (20-30 minute)
migration that should probably be re-written in SQL.