Bug/inline image opt (#1231)
* Change image opt tests to use an accepted image source URI

Because we are about to change the image opt implementation to only
operate on accepted image sources.  Otherwise, this commit makes no
functional change.

* Add the "IMOPT_ALLOWED_SOURCES" config value

The value is an array of whitelisted image sources.  The image optimization code
is about to be changed to pass through, unchanged, any image URLs from unknown sources.

* Don't optimize image URLs having unknown hostnames [ch5642]

* Make a test work all hours of the day